Move dataype definitions to their own class
